requirements.txt
absl-py==0.9.0
asgiref==3.2.3
asn1crypto==1.2.0
astor==0.7.1
autopep8==1.4.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
ConfigArgParse==0.13.0
cryptography==2.8
Cython==0.29.14
Django==3.0.2
django-filter==2.2.0
django-nested-admin==3.2.4
django-rest-knox==4.1.0
djangorestframework==3.11.0
ecdsa==0.13
Flask==1.1.1
Flask-Cors==3.0.8
flask-marshmallow==0.10.1
future==0.18.2
gast==0.2.2
google-pasta==0.1.8
grpcio==1.23.0
h5py==2.8.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
mxnet==1.5.0
numpy==1.18.0
olefile==0.46
opencv-python==4.1.2.30
opt-einsum==0+untagged.49.gdbede45.dirty
Pillow==7.0.0
pipenv==2018.11.26
protobuf==3.11.2
psycopg2==2.8.4
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.9.4
Pygments==2.5.2
pyOpenSSL==19.1.0
PySocks==1.7.1
python-jose-cryptodome==1.3.2
python-monkey-business==1.0.0
pytz==2019.3
pyzmq==18.1.1
requests==2.22.0
scipy==1.4.1
six==1.13.0
SQLAlchemy==1.3.11
sqlparse==0.3.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
urllib3==1.25.7
virtualenv==16.7.9
virtualenv-clone==0.5.3
Werkzeug==0.16.0
wrapt==1.11.2
